Trait sequence_trie::TrieKey

pub trait TrieKey: Eq + Hash { }

Aggregate trait for types which can be used to key a SequenceTrie.

This trait is automatically implemented for all types implementing the supertraits.