Crate sequence_trie

Sequence Trie - a trie-like data-structure for storing sequences of values.

See the SequenceTrie type for documentation.

Structs

Iter
Iterator over the keys and values of a SequenceTrie.

Keys
Iterator over the keys of a SequenceTrie.

PrefixIter
Iterator over the longest prefix of nodes which matches a key.

SequenceTrie
A SequenceTrie is recursively defined as a value and a map containing child Tries.

Values
Iterator over the values of a SequenceTrie.

Traits

TrieKey
Aggregate trait for types which can be used to key a SequenceTrie.

Type Definitions

KeyValuePair
Vector of key fragment references and values, yielded during iteration.